refact(users-context): moving users state management logic to its own zustand object
